Generalized URL parsing in the agent using the URL parser from the agent adapter.
