+++ src/mtconnect/agent.hpp
@@ -279,8 +279,7 @@ namespace mtconnect {
     /// @param[in] deviceXml the device xml as a string
     /// @param[in] source the source loading the device
     void loadDevices(std::list<DevicePtr> device,
-                     const std::optional<std::string> source = std::nullopt,
-                     bool force = false);
+                     const std::optional<std::string> source = std::nullopt, bool force = false);
 
     /// @brief receive and parse a single device from a source
     /// @param[in] deviceXml the device xml as a string
@@ -591,8 +590,10 @@ namespace mtconnect {
     void deliverConnectStatus(entity::EntityPtr, const StringList &devices,
                               bool autoAvailable) override;
     void deliverCommand(entity::EntityPtr) override;
-    void deliverDevice(DevicePtr device) override { m_agent->loadDevices({device}, std::nullopt,
-                                                                         true); }
+    void deliverDevice(DevicePtr device) override
+    {
+      m_agent->loadDevices({device}, std::nullopt, true);
+    }
     void deliverDevices(std::list<DevicePtr> devices) override { m_agent->loadDevices(devices); }
 
     void sourceFailed(const std::string &identity) override { m_agent->sourceFailed(identity); }
