+++ src/mtconnect/asset/asset.hpp
@@ -160,9 +160,9 @@ namespace mtconnect {
       /// Override to skip the `hash`, `timestamp`, and `removed` properties.
       ///
       /// @param[in,out] sha1 The boost sha1 accumulator
-      void hash(boost::uuids::detail::sha1 &sha1) const override
+      void hash(::boost::uuids::detail::sha1 &sha1) const override
       {
-        static const boost::unordered_set<std::string> skip {"hash", "timestamp", "removed"};
+        static const ::boost::unordered_set<std::string> skip {"hash", "timestamp", "removed"};
         entity::Entity::hash(sha1, skip);
       }
