+++ src/mtconnect/source/adapter/agent_adapter/session.hpp
@@ -24,7 +24,6 @@
 
 #include "mtconnect/config.hpp"
 #include "mtconnect/entity/entity.hpp"
-#include "url_parser.hpp"
 
 namespace mtconnect::source::adapter {
   struct Handler;
@@ -53,7 +52,7 @@ namespace mtconnect::source::adapter::agent_adapter {
       /// @param stream `true` if HTTP x-multipart-replace streaming is desired
       /// @param next Function to determine what to do on successful read
       Request(const std::optional<std::string> &device, const std::string &operation,
-              const UrlQuery &query, bool stream, Next next)
+              const url::UrlQuery &query, bool stream, Next next)
         : m_sourceDevice(device),
           m_operation(operation),
           m_query(query),
@@ -65,7 +64,7 @@ namespace mtconnect::source::adapter::agent_adapter {
 
       std::optional<std::string> m_sourceDevice;  ///< optional source device
       std::string m_operation;     ///< The REST operation (probe, current, sample, asset)
-      UrlQuery m_query;            ///< URL Query parameters
+      url::UrlQuery m_query;       ///< URL Query parameters
       bool m_stream;               ///< `true` if using HTTP long pull
       Next m_next;                 ///< function to call on successful read
       int32_t m_agentVersion = 0;  ///< agent version if required > 0 for asset requests
@@ -73,7 +72,10 @@ namespace mtconnect::source::adapter::agent_adapter {
       /// @brief Given a url, get a formatted target for a given operation
       /// @param url The base url
       /// @return a string with a new URL path and query (for the GET)
-      auto getTarget(const Url &url) { return url.getTarget(m_sourceDevice, m_operation, m_query); }
+      auto getTarget(const url::Url &url)
+      {
+        return url.getTarget(m_sourceDevice, m_operation, m_query);
+      }
     };
 
     virtual ~Session() {}
